+++ src/features/navigation/controls/index.ts
@@ -24,8 +24,8 @@ const useGetMapControls = (): Collection<Control> | Control[] | undefined => {
             placeholder: t("search_engine_placeholder"),
         }),
         new ScaleLine(),
-        new GeoportalZoom({ position: "bottom-right", zoomInTipLabel: "Zoom In", zoomOutTipLabel: "Zoom Out" }),
         new GeoportalFullScreen({ position: "bottom-right", tipLabel: t("full_screen_label") }),
+        new GeoportalZoom({ position: "bottom-right", zoomInTipLabel: "Zoom In", zoomOutTipLabel: "Zoom Out" }),
         catalogControl,
         centerToReportControl,
     ];
